+++ python/lsst/ts/mtdomegui/tab/tab_aperture_shutter.py
@@ -144,7 +144,7 @@ class TabApertureShutter(TabTemplate):
             "drive_torque": TabFigure(
                 "Actual Drive Torque",
                 self.model,
-                "J",
+                "N*m",
                 [str(idx) for idx in range(NUM_DRIVE_SHUTTER)],
             ),
             "drive_current": TabFigure(
@@ -361,18 +361,18 @@ class TabApertureShutter(TabTemplate):
         position_actual = telemetry["positionActual"]
         for idx in range(APSCS_NUM_SHUTTERS):
             self._status["position_commanded"][idx].setText(
-                f"{position_commanded[idx]:.2f} deg"
+                f"{position_commanded[idx]:.2f} %"
             )
             self._status["position_actual"][idx].setText(
-                f"{position_actual[idx]:.2f} deg"
+                f"{position_actual[idx]:.2f} %"
             )
 
         for idx in range(NUM_DRIVE_SHUTTER):
             self._status["drive_torque_commanded"][idx].setText(
-                f"{telemetry['driveTorqueCommanded'][idx]:.2f} J"
+                f"{telemetry['driveTorqueCommanded'][idx]:.2f} N*m"
             )
             self._status["drive_torque_actual"][idx].setText(
-                f"{telemetry['driveTorqueActual'][idx]:.2f} J"
+                f"{telemetry['driveTorqueActual'][idx]:.2f} N*m"
             )
             self._status["drive_current_actual"][idx].setText(
                 f"{telemetry['driveCurrentActual'][idx]:.2f} A"