+++ python/lsst/ts/mtdomegui/tab/tab_azimuth.py
@@ -153,7 +153,7 @@ class TabAzimuth(TabTemplate):
             "drive_torque": TabFigure(
                 "Actual Drive Torque",
                 self.model,
-                "J",
+                "N*m",
                 [str(idx) for idx in range(AMCS_NUM_MOTORS)],
             ),
             "drive_current": TabFigure(
@@ -382,10 +382,10 @@ class TabAzimuth(TabTemplate):
 
         for idx in range(AMCS_NUM_MOTORS):
             self._status["drive_torque_commanded"][idx].setText(
-                f"{telemetry['driveTorqueCommanded'][idx]:.2f} J"
+                f"{telemetry['driveTorqueCommanded'][idx]:.2f} N*m"
             )
             self._status["drive_torque_actual"][idx].setText(
-                f"{telemetry['driveTorqueActual'][idx]:.2f} J"
+                f"{telemetry['driveTorqueActual'][idx]:.2f} N*m"
             )
             self._status["drive_current_actual"][idx].setText(
                 f"{telemetry['driveCurrentActual'][idx]:.2f} A"