+++ doc/version_history.rst
@@ -6,6 +6,16 @@
 Version History
 ##################
 
+.. _lsst.ts.mtdomegui-0.4.3:
+
+-------------
+0.4.3
+-------------
+
+* Allow to query the shutter status.
+* Fix the unit of position of the shutter.
+* Change the torque unit to be N*m instead of J.
+
 .. _lsst.ts.mtdomegui-0.4.2:
 
 -------------